moving admin-tools lib to lib module
